make react-server-dom-webpack plugin reads file path from "result.path"
it fixes an error happen when the plugin tries to read a file with a hashtag (#) character in its path
The hashtag character is escaped by adding `\x00` character before it. So, the second parameter in `normalResolver.resolve` returns the escaped version of the path
when the plugin tries to read its content, it gets the error
```
TypeError [ERR_INVALID_ARG_VALUE]: The argument 'path' must be a string or Uint8Array without null bytes. Received '/path/to/file/with/\x00#in/it'
```